Star streaks easter egg at rotational.io/starstreak/
This PR adds the original PS5 star streaks javascript back to the
website at an easter egg page for posterity and in case anyone is
feeling a bit nostalgic. Adding this page also allows us to delete the
old repository so there is no confusion about which repo is our website.