Complete draft of can't read won't buy (#13)
* complete draft of can't read won't buy

* fix image width and stray blockquote

* fixed some typos

* Update 2021-04-27-cant-read-wont-buy.md

Great post, @rebeccabilbro ! I made light edits, added a sentence on psychological safety, and strengthened the ending. I also double-checked that draft = false!

A comment on writing style: you are both excellent writers. However, you both use a lot of what I call "hedging" language e.g. "seems to be", "might be", "could be", "may be" etc. This is common in academia so I'm not surprised to see it. In a business context, I think it's important to be direct and take a position/ stance as opposed to hedging. So stating what is/ we believe vs what could be/ might be.

Co-authored-by: Benjamin Bengfort <benjamin@bengfort.com>
Co-authored-by: Edwin Schmierer <edwin@rotational.io>