Hello World Blog Post (#25)
* Hello World Blog Post

This PR creates a Hello World blog post post-dated to our founding date
on January 21, 2021. Because we already had a blog on that date, I
shifted its date to February 12, 2021. A bit of history rewriting for
the sake of clarity and narrative structure that I hope won't be too
egregious to anyone reading this PR or commit message.

* add complexity through simple components

* Update 2021-01-21-hello-world.md

Perfect! Just fixed a typo..

Co-authored-by: Edwin Schmierer <edwin@rotational.io>